requirements.txt
isort>=5.10.1
python-dateutil>=2.8.2
requests>=2.28.1
tabulate>=0.9.0
